Initiate articles collection - PII, PMID, title
/initiate_journal_db/:journalname
———————————————
XML versioning is based on PII. So when crawling for XML and updating
the DB, if there is no PII in the XML we cannot update the DB or upload
the XML.

this path will use the current production DB and PubMed to match PII to
PMID via title. Articles collection gets created and ids.pii, ids.pmid,
and title get inserted. If there are ANY docs in the articles
collection then this path will NOT insert into the DB